- systemd-sysvinit creates a symlink from '/bin/systemd' to   '/sbin/init'. Unfortunately kiwi linuxrc tries to check   for the file permissions which leads to a false positive   rebootException (bnc #700603)

+++ modules/KIWILinuxRC.sh
@@ -5031,7 +5031,7 @@ function validateRootTree {
 	# check whether that mount is a valid system tree or not. Therefore
 	# some sanity checks are made here
 	# ----
-	if [ ! -x /mnt/sbin/init ];then
+	if [ ! -x /mnt/sbin/init -a ! -L /mnt/sbin/init ];then
 		systemException "/sbin/init no such file or not executable" "reboot"
 	fi
 }
